refactor: centralize cashflow timeline utility
